Selection Service - Refactoring
