+++ projects/annu-ng-lib/src/lib/components/content-editor/content-editor.interface.ts
@@ -20,9 +20,3 @@ export interface EditorElement {
     data?: EditorElementData;
     children?: Array<EditorElement>;
 };
-
-export interface TextSelectionEvent {
-    text: string;
-    selection: Selection;
-    selectionRect: DOMRect;
-}
