+++ projects/annu-ng-lib/src/lib/components/content-editor/heading/heading.component.scss
@@ -1,0 +1,3 @@
+h1, h2, h3, h4, h5, h6 {
+    position: relative;
+}
