+++ projects/annu-ng-lib/src/lib/components/content-editor/list-item/list-item.component.scss
@@ -1,0 +1,3 @@
+li {
+    position: relative;
+}
