+++ projects/annu-ng-lib/src/lib/components/content-editor/list-item/list-item.component.ts
@@ -40,4 +40,8 @@ export class ListItemComponent implements OnInit {
     event.stopPropagation();
     this.focusin.emit(this.value);
   }
+
+  public textSelected() {
+    // TODO
+  }
 }
