+++ projects/annu-ng-lib/src/lib/components/content-editor/paragraph/paragraph.component.html
@@ -1,3 +1,3 @@
-<p contenteditable [(ngModel)]="value.data.text" (keydown.enter)="enterPressed($event)"
+<p contenteditable [(ngModel)]="value.data.text" (keydown.enter)="enterPressed($event)" (anuFormatInline)="textSelected()"
     (keydown.backspace)="backspacePressed($event)" [name]="value.name" [focus]="value?.focused" (focusin)="onFocus($event)"
     (blur)="onBlur($event)"></p>
