+++ projects/annu-ng-lib/src/lib/components/content-editor/paragraph/paragraph.component.scss
@@ -1,0 +1,3 @@
+p {
+    position: relative;
+}
