+++ projects/annu-ng-lib/src/lib/components/content-editor/paragraph/paragraph.component.ts
@@ -40,4 +40,8 @@ export class ParagraphComponent implements OnInit {
     event.stopPropagation();
     this.focusin.emit(this.value);    
   }
+
+  public textSelected() {
+    // TODO
+  }
 }
