+++ projects/annu-ng-lib/src/lib/components/content-editor/services/selection.service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+
+import { SelectionService } from './selection.service';
+
+describe('SelectionService', () => {
+  let service: SelectionService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(SelectionService);
+  });
+
+  it('should be created', () => {
+    expect(service).toBeTruthy();
+  });
+});
